requirements.txt
asgiref==3.9.1
Django==4.2.23
sqlparse==0.5.3
typing_extensions==4.14.1
djangorestframework==3.14.0
drf-yasg==1.21.6
djangorestframework-simplejwt==5.3.0
